[Edit] 회원가입 및 로그인 기능 수정
